+++ src/base/cs_parameters_check.c
@@ -1619,6 +1619,14 @@ cs_parameters_check(void)
   }
 
   /* Stokes */
+
+  /* transposed velocity gradient term and secondary viscosity */
+  cs_parameters_is_in_range_int(CS_ABORT_DELAYED,
+                                _("while reading input data"),
+                                "cs_glob_stokes_model->ivisse",
+                                cs_glob_stokes_model->ivisse,
+                                0, 2);
+
   cs_parameters_is_in_range_int(CS_ABORT_DELAYED,
                                 _("while reading input data"),
                                 "cs_glob_stokes_model->iprco",
@@ -1659,13 +1667,11 @@ cs_parameters_check(void)
                                 cs_glob_stokes_model->idilat,
                                 0, 6);
 
-  if (cs_glob_stokes_model->idilat == 0) {
-    cs_parameters_is_greater_int(CS_ABORT_DELAYED,
-                                  _("while reading input data"),
-                                  "Boussinesq approximation requires the thermal_expansion field",
-                                  cs_field_by_name("thermal_expansion"),
-                                  0);
-  }
+  cs_parameters_is_in_range_int(CS_ABORT_DELAYED,
+                                _("while reading input data"),
+                                "cs_glob_stokes_model->iphydr",
+                                cs_glob_stokes_model->iphydr,
+                                0, 3);
 
   cs_field_get_key_struct(CS_F_(vel), key_cal_opt_id, &var_cal_opt);
   /* steady or variable time step time algorithm not compatible with theta